fix the way we check for Pima's percent data in the data layers plugin
This is clearly not the way this should be handled, but the plugin was hardcoded to look for percentage data on a block group layer to begin with. It now checks if the module is Pima County, in which case it instead looks for percentage data on a precinct layer.