create Pima AZ module and event page (#655)
* create Pima AZ module and event page

* hardcode Pima County into addTracts as a precinct-level exception

* fix Pima County boundary geojsons

* fix the way we check for Pima's percent data in the data layers plugin

This is clearly not the way this should be handled, but the plugin was hardcoded to look for percentage data on a block group layer to begin with. It now checks if the module is Pima County, in which case it instead looks for percentage data on a precinct layer. This also removes the Pima check from addTracts in index.js, as it was unnecessary.

* edit Pima event page text